Updates for eckit-1.1.0 (#33)
* Update eckit 0.23 -> 1.1.0

* Added DOWNLOAD_ONLY option

* Use .r2 for odb-api build

* Updated to version r2 of odb api tar ball.

+++ buildscripts/libs/build_eccodes.sh
@@ -44,6 +44,7 @@ cd ${JEDI_STACK_ROOT}/${PKGDIR:-"pkg"}
 
 software=$name-$version
 [[ -d $software ]] || ( git clone -b $version $gitURL $software )
+[[ ${DOWNLOAD_ONLY} =~ [yYtT] ]] && exit 0
 [[ -d $software ]] && cd $software || ( echo "$software does not exist, ABORT!"; exit 1 )
 [[ -d build ]] && rm -rf build
 mkdir -p build && cd build
