+++ buildscripts/libs/build_pnetcdf.sh
@@ -43,6 +43,7 @@ cd ${JEDI_STACK_ROOT}/${PKGDIR:-"pkg"}
 software=$name-$version
 url="https://parallel-netcdf.github.io/Release/$software.tar.gz"
 [[ -d $software ]] || ( wget $url; tar -xf $software.tar.gz )
+[[ ${DOWNLOAD_ONLY} =~ [yYtT] ]] && exit 0
 [[ -d $software ]] && cd $software || ( echo "$software does not exist, ABORT!"; exit 1 )
 [[ -d build ]] && rm -rf build
 mkdir -p build && cd build
