+++ buildscripts/libs/build_xerces.sh
@@ -40,6 +40,7 @@ software=$name-c-$version
 vnum=$(echo $version | cut -d. -f1)
 url=https://archive.apache.org/dist/xerces/c/$vnum/sources/$software.tar.gz
 [[ -d $software ]] || ( wget $url; tar -xf $software.tar.gz )
+[[ ${DOWNLOAD_ONLY} =~ [yYtT] ]] && exit 0
 [[ -d $software ]] && cd $software || ( echo "$software does not exist, ABORT!"; exit 1 )
 [[ -d build ]] && rm -rf build
 mkdir -p build && cd build
